+++ scratchpad/WebGL/webgl-scratchpad/src/shaders/particles_f.glsl
@@ -1,5 +1,6 @@
 #version 300 es
-precision highp float;
+
+precision mediump float;
 
 uniform sampler2D uSampler;
 uniform vec3 uColor;
